+++ extensions/monobank/CHANGELOG.md
@@ -1,5 +1,9 @@
 # monobank Changelog
 
+## [Fix] - 2025-12-02
+
+- Used country flags in .png format so they look the same on both macOS and Windows platforms
+
 ## [Fix] - 2025-11-27
 
 - Added Windows support